pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pymysql>=1.1.2",
]
